* 'master' of https://github.com/StevoGTA/CppToolbox:
  Windows - Renamed CMediaFoundationVideoCodec => CMediaFoundationDecodeVideoCodec. Windows - Updated CDirectXTexture and CDirectXGPU to latest methods. Windows - Added applicationFolder to CFolder. Windows - Removed warnings in CString-WIndows. Windows - Removed warnings in CMediaFoundationServices. Windows - Updated CFileDataSource-WIndows-WinRT. Removed warnings in CImage. Updated Windows notes on building CImage libraries.
  Finalized DeferredNotificationCenter for Windows. Fixed issue in CThread::getCurrentRef(). Removed warnings CString-Windows and CMediaFoundationServices.

+++ Source/Add On - Windows/CString-Windows.cpp
@@ -538,9 +538,9 @@ TBuffer<UTF32Char> CString::getUTF32Chars() const
 //----------------------------------------------------------------------------------------------------------------------
 {
 	// Collect characters
-	TBuffer<UTF32Char>	buffer(mString.length());
-	::WideCharToMultiByte(sGetCodePageForCStringEncoding(kEncodingUTF32Native), 0, mString.c_str(), mString.length(),
-			(char*) *buffer, mString.length() * 4, NULL, NULL);
+	TBuffer<UTF32Char>	buffer((UInt32) mString.length());
+	::WideCharToMultiByte(sGetCodePageForCStringEncoding(kEncodingUTF32Native), 0, mString.c_str(),
+			(int) mString.length(), (char*) *buffer, (int) mString.length() * 4, NULL, NULL);
 
 	return buffer;
 }
